+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this AWS Lambda@Edge project will be documented in this f
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.3] - December 08, 2023
+
+### Changed
+
+- Bump @kameleoon/nodejs-sdk version to 3.2.0
+- Change the value of `kameleoon_user_id` to `kameleoonVisitorCode`
+
 ## [1.0.2] - September 26, 2023
 
 ### Changed